Refactor this to another name, because it will be common logic later.
This isn't specifically getting PID children, it's just walking graph edges in a
map.

Right now, the map is built pid -> pidChildren. We can easily reverse that, and
create a map of pid -> pidParent, and apply this function to said map, to get
the process parents as well as the children.